move utils to data pkg
